moved map creator into its own class because we will need it on other activities too
